bootcamp_cloud/projetos/lambda-api-rds-dynamo/pyproject.toml
[tool.poetry]
name = "lambda-api-rds-dynamo"
version = "0.1.0"
description = ""
authors = ["Lucas-Sobreira <lucas-andrade2011@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
streamlit = "^1.45.0"
sqlalchemy = "^2.0.40"
pandas = "^2.2.3"
plotly = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
